# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FastStart API",
    "description": "Registry for reusable code snippets with email and embeddings utilities",
    "version": "0.0.0",
    "contact": {
      "name": "Enfiniq",
      "url": "https://github.com/Enfiniq/FastStart"
    }
  },
  "servers": [
    {
      "url": "https://faststart.neploom.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "FastStartKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Package-specific authentication key"
      },
      "ScopeSpecificKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scope-specific authentication key"
      },
      "AdminKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Administrative access key"
      }
    },
    "schemas": {
...
      "FastStartRecord": {
        "type": "object",
        "required": ["name", "fastStart", "author", "type"],
        "properties": {
          "id": { "type": "integer", "description": "Database ID" },
          "name": { "type": "string" },
          "fastStart": { "type": "string" },
          "author": { "type": "string" },
          "type": { "type": "string" },
          "versions": {
            "type": "array",
            "items": { "type": "string" },
            "default": ["1.0.0"]
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "FastStartRecordInput": {
        "type": "object",
        "required": ["name", "fastStart", "author", "type"],
        "properties": {
          "name": { "type": "string" },
          "fastStart": { "type": "string" },
          "author": { "type": "string" },
          "type": { "type": "string" },
          "versions": {
            "type": "array",
            "items": { "type": "string" },
            "default": ["1.0.0"]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/database": {
      "get": {
        "tags": ["Database"],
        "summary": "List packages from database",
        "description": "Lists packages from PostgreSQL database with filtering and pagination",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by package name (partial match)"
          },
          {
            "name": "author",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by author name (partial match)"
          },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by package type (exact match)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "description": "Number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "minimum": 0, "default": 0 },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Packages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FastStartRecord"
                      }
                    },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["Database"],
        "summary": "Sync package to GitHub",
        "description": "Syncs a package from database to GitHub registry",
        "security": [{ "FastStartKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Package synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Package posted to GitHub registry"
                    },
                    "package": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "put": {
        "tags": ["Database"],
        "summary": "Create or update package",
        "description": "Creates or updates a package in the database",
        "security": [{ "FastStartKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FastStartRecordInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Package created/updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Package created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/FastStartRecord"
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["Database"],
        "summary": "Remove package",
        "description": "Removes package from both database and GitHub registry",
        "security": [{ "FastStartKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Package removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Package deleted from both database and GitHub registry"
                    },
                    "package": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
...
}